requirements.txt
absl-py==2.0.0
astunparse==1.6.3
boto3==1.33.13
botocore==1.33.13
cachetools==5.3.2
captum==0.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cycler==0.11.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.38.0
fsspec==2023.1.0
gast==0.4.0
google-auth==2.25.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==6.7.0
jiwer==3.0.3
jmespath==1.0.1
joblib==1.3.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
nltk==3.8.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==9.5.0
protobuf==3.20.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
rapidfuzz==3.4.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.8.2
safetensors==0.4.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
transformers==4.30.2
typing_extensions==4.7.1
urllib3==1.26.18
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
